+++ functions/ListOutput.fnc.php
@@ -494,7 +494,7 @@ function ListOutput( $result, $column_names, $singular = '.', $plural = '.', $li
 						issetVal( $link['add']['title'], '' ),
 						( mb_strpos( $link['add']['link'], '"' ) === 0 ?
 							$link['add']['link'] :
-							'"' . URLEscape( $link['add']['link'] ) . '"' )
+							URLEscape( $link['add']['link'] ) )
 					) . '</td></tr>';
 			}
 			elseif ( $link['add']['span'] && ! isset( $_REQUEST['_ROSARIO_PDF'] ) )
@@ -632,7 +632,7 @@ function ListOutput( $result, $column_names, $singular = '.', $plural = '.', $li
 					issetVal( $link['add']['title'], '' ),
 					( mb_strpos( $link['add']['link'], '"' ) === 0 ?
 						$link['add']['link'] :
-						'"' . URLEscape( $link['add']['link'] ) . '"' )
+						URLEscape( $link['add']['link'] ) )
 				) . '</td></tr>';
 			}
 			elseif ( isset( $link['add']['span'] ) && ! isset( $_REQUEST['_ROSARIO_PDF'] ) )
@@ -683,7 +683,7 @@ function ListOutput( $result, $column_names, $singular = '.', $plural = '.', $li
 				issetVal( $link['add']['title'], '' ),
 				( mb_strpos( $link['add']['link'], '"' ) === 0 ?
 					$link['add']['link'] :
-					'"' . URLEscape( $link['add']['link'] ) . '"' )
+					URLEscape( $link['add']['link'] ) )
 			) . '</div>';
 		}
 		elseif (  ( ! empty( $link['add']['html'] )