+++ modules/Discipline/DisciplineForm.php
@@ -451,13 +451,13 @@ function _makeRemove( $value, $column )
 		{
 			$return = button(
 				'remove', _( 'Don\'t use' ),
-				'"' . URLEscape( 'Modules.php?modname=' . $_REQUEST['modname'] . '&modfunc=delete_usage&id=' . $THIS_RET['USAGE_ID'] ) . '"'
+				URLEscape( 'Modules.php?modname=' . $_REQUEST['modname'] . '&modfunc=delete_usage&id=' . $THIS_RET['USAGE_ID'] )
 			);
 
 			$return .= '<br />' . button(
 				'remove',
 				_( 'Delete' ),
-				'"' . URLEscape( 'Modules.php?modname=' . $_REQUEST['modname'] . '&modfunc=delete&id=' . $THIS_RET['ID'] ) . '"'
+				URLEscape( 'Modules.php?modname=' . $_REQUEST['modname'] . '&modfunc=delete&id=' . $THIS_RET['ID'] )
 			);
 		}
 		else
@@ -465,7 +465,7 @@ function _makeRemove( $value, $column )
 			$return = button(
 				'add',
 				_( 'Use at this school' ),
-				'"' . URLEscape( 'Modules.php?modname=' . $_REQUEST['modname'] . '&modfunc=add_usage&id=' . $THIS_RET['ID'] ) . '"'
+				URLEscape( 'Modules.php?modname=' . $_REQUEST['modname'] . '&modfunc=add_usage&id=' . $THIS_RET['ID'] )
 			);
 		}
 	}