+++ modules/School_Setup/GradeLevels.php
@@ -248,5 +248,5 @@ function _makeRemoveButton( $value, $column )
 	$button_link = 'Modules.php?modname=' . $_REQUEST['modname'] . '&modfunc=remove&id=' .
 		$THIS_RET['ID'];
 
-	return button( 'remove', '', '"' . URLEscape( $button_link ) . '"' );
+	return button( 'remove', '', URLEscape( $button_link ) );
 }