+++ modules/School_Setup/Periods.php
@@ -341,5 +341,5 @@ function _makeRemoveButton( $value, $column )
 	$button_link = 'Modules.php?modname=' . $_REQUEST['modname'] . '&modfunc=remove&id=' .
 		$THIS_RET['PERIOD_ID'];
 
-	return button( 'remove', '', '"' . URLEscape( $button_link ) . '"' );
+	return button( 'remove', '', URLEscape( $button_link ) );
 }