+++ modules/Students/includes/Address.inc.php
@@ -741,10 +741,10 @@ if ( ! $_REQUEST['modfunc'] )
 					$remove_button = button(
 						'remove',
 						'',
-						'"' . URLEscape( 'Modules.php?modname=' . $_REQUEST['modname'] .
+						URLEscape( 'Modules.php?modname=' . $_REQUEST['modname'] .
 						'&category_id=' . $_REQUEST['category_id'] .
 						'&modfunc=delete_address&address_id=' . $_REQUEST['address_id'] .
-						'&person_id=' . $contact['PERSON_ID'] ) . '"'
+						'&person_id=' . $contact['PERSON_ID'] )
 					);
 				}
 
@@ -1262,11 +1262,11 @@ if ( ! $_REQUEST['modfunc'] )
 							echo '<td>' . button(
 								'remove',
 								'',
-								'"' . URLEscape( 'Modules.php?modname=' . $_REQUEST['modname'] .
+								URLEscape( 'Modules.php?modname=' . $_REQUEST['modname'] .
 								'&category_id=' . $_REQUEST['category_id'] .
 								'&modfunc=delete_address&address_id=' . $_REQUEST['address_id'] .
 								'&person_id=' . $_REQUEST['person_id'] .
-								'&contact_id=' . $info['ID'] ) . '"'
+								'&contact_id=' . $info['ID'] )
 							) . '</td><td>';
 						}
 						else